pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "age-untick"
version = "0.0.0"
dynamic = ["version"]
dependencies = [
    "google-cloud-vision",
    "google-cloud-storage",
    "google-generativeai",
    "pillow"
]

[project.optional-dependencies]
dev = [
    "setuptools",
    "python-dotenv"
]
